Provide session_id header
The session_id header is always available and is fixed for every user.
This header ensures that we can identify the user across multiple
services.  Ideally, the session will be dropped in the triplestore to
ensure all users know what there is to know about this particular
user (which means: the session should become a URI).

There is a huge hickup with the current approach though.  The random
operator for the session id isn't exactly random.  We should set the
seed so it's always different.