Use UUID as user identifier
We can define our own random string in order to identify a connection,
but we can do better and use someone's proven UUID library instead.  The
latter is what we are doing here.  We use uuid1 because we don't really
know much better than that.  It seems to be random when tested with mix
so it's probably good enough.