Assume clean targets
A clean target would bind to port 80 and have a clean name.  The
previous solution assumed that we would be running the identifier
manually, this is not hte case anymore.  For this reason, we're
switching to a named dispatcher running on port 80 and to our own
service running on port 80 as well.