Add Access-Control-Allow-Origin configuration through MU_CORS_HEADER
The MU_CORS_HEADER environment variable sets the
Access-Control-Allow-Origin header.  This header 'fixes' the CORS
issue hence it will be easiest to find under this name.

Advised use for public services is `MU_CORS_HEADER: "*"`, which will
add the header `Access-Control-Allow-Origin: "*"` header to responses
for which the backend microservice did not set such a header.  Not
setting the environment variable means the header should not be set by
the identifier.