Support for default MU_AUTH_ALLOWED_GROUPS
When a user arrives at the stack for the first time, there is no
session available and there are no MU_AUTH_ALLOWED_GROUPS.  This means
the cache can't respond to these requests.  With this change, the
MU_AUTH_ALLOWED_GROUPS can be set for sessions in which none is known
yet.

This approach makes cache hits much more viable for public users and
may result in more complete responses to be constructed for first-time
viewers and bots.