Don't use default when CLEAR header has just been sent
The CLEAR header is currently the way in which login services indicate
the header should be cleared out and get recalculated.  As there is
currently no more in-depth way of informing the identifier, we are
currently assuming this CLEAR header has to be reused in the next
call.  For this reason, we set the CLEAR header as the access rights
on the request, and don't set any access rights header on the next
requests, untill a request supplies us with a new set of
mu-auth-allowed-groups.