Compress responses when possible
Cowboy can compress responses.  This should be done somewhere in the
stack.  Either we should give this responsability to the backend
microservices (which means they should all care about it) or we should
treat it in a single place.  The mu-identifier seems to be the best
place for now as it's a single point and is easily controlled.