Removing mix lock file for now
The mix lock file serves its purpose when upstream builds are fixed.
In the current setting we're keeping them open for a while to easily
test stability of upstream updates on the go.  We depend on Docker for
the release management, making sure we can replicate the setup when
needed.