Add observer_cli and fix plug_mint_proxy version
The version of plug_mint_proxy was not fixed.  This is now still being
pulled from GitHub, but it is a tag so it's a fixed version.

The observer_cli was not in the system.  We have added it because it
gives us some insights into the most heavy processes over a terminal.
A sane devops addition which will help us identify issues before they
broak hard.