Support for overriding vary header
The vary header allows to configure when to cache things and when not to
cache things.  This is an import part for dispatching on
request-type (for example).  Because various parts of the stack emit
various vary headers, it's hard for browsers to figure out what to do
when.

This commit adds an override to the vary header so you can control it in
a single location and get some more sanity into the approach.  It's
experimental, it may go away, and it may well be that we find better
approaches to handle this in different parts of the stack.  For now,
this may help to solve practical cases.