feat: make Cowboy's idle_timeout configurable
This is useful when the identifier is part of a stack that will serve
large files. Downloading said files will take a while, and if it takes
more than a minute (the default timeout) Cowboy will close ongoing
requests, resulting in partially downloaded files and unhappy users.

See this Cowboy issue for some context:
https://github.com/ninenines/cowboy/issues/1502